fix: test failures, derive tuple structs, replace manual impls
- TxEncodingFormat: replace WordLayout with TryFrom<u32>, return error
  instead of panicking on invalid oracle input
- B160::read_words: assert upper 32 bits are zero
- WordLayout derive macro: add tuple struct support
- Derive WordLayout for BlockHashes, DelegatedU256, KZGCommitmentAndProof
  (replacing manual impls)
- Fix unit tests to match new validation approach

+++ forward_system/src/run/query_processors/tx_data.rs
@@ -105,7 +105,7 @@ impl<TS: TxSource> OracleQueryProcessor for TxDataResponder<TS> {
                     );
                 };
 
-                format.write_words(&mut |w| result.push(w));
+                (format as u32).write_words(&mut |w| result.push(w));
             }
             TX_FROM_QUERY_ID => {
                 let Some(from) = self.next_tx_from.take() else {
